Split client forms code base into different components
