+++ app/javascript/src/components/Clients/ClientForm/index.tsx
@@ -4,46 +4,17 @@ import React, { useEffect, useState } from "react";
 import "react-phone-number-input/style.css"; //eslint-disable-line
 import { Country, State, City } from "country-state-city";
 import { Formik, Form, FormikProps } from "formik";
-import { EditImageButtonSVG, deleteImageIcon } from "miruIcons";
 import PhoneInput from "react-phone-number-input";
 import flags from "react-phone-number-input/flags";
-import { Avatar, Toastr } from "StyledComponents";
+import { Toastr } from "StyledComponents";
 
 import clientApi from "apis/clients";
 import CustomReactSelect from "common/CustomReactSelect";
 import { InputErrors, InputField } from "common/FormikFields";
 
 import { clientSchema, getInitialvalues } from "./formValidationSchema";
+import UploadLogo from "./UploadLogo";
 import { formatFormData } from "./utils";
-
-import { i18n } from "../../../i18n";
-
-interface IClientForm {
-  clientLogoUrl: string;
-  handleDeleteLogo: any;
-  setClientLogoUrl: any;
-  setClientLogo: any;
-  formType?: string;
-  clientData?: any;
-  setClientData?: any;
-  setnewClient?: any;
-  clientLogo?: any;
-  setApiError?: any;
-  setShowEditDialog?: any;
-}
-
-interface FormValues {
-  name: string;
-  email: string;
-  phone: string;
-  address1: string;
-  address2: string;
-  country: any;
-  state: any;
-  city: any;
-  zipcode: string;
-  logo: any;
-}
 
 const ClientForm = ({
   clientLogoUrl,
@@ -101,38 +72,6 @@ const ClientForm = ({
     return cities;
   };
 
-  const onLogoChange = e => {
-    const file = e.target.files[0];
-    const isValid = isValidFileUploaded(file);
-
-    if (isValid.fileExtension && isValid.fileSizeValid) {
-      setClientLogoUrl(URL.createObjectURL(file));
-      setClientLogo(file);
-    } else {
-      if (!isValid.fileExtension && !isValid.fileSizeValid) {
-        setFileUploadError(
-          i18n.t("invalidImageFormatSize", { fileSize: "30" })
-        );
-      } else if (isValid.fileExtension && !isValid.fileSizeValid) {
-        setFileUploadError(i18n.t("invalidImageSize", { fileSize: "30" }));
-      } else {
-        setFileUploadError(i18n.t("invalidImageFormat"));
-      }
-    }
-  };
-
-  const isValidFileUploaded = file => {
-    const validExtensions = ["png", "jpeg", "jpg"];
-    const fileExtensions = file.type.split("/")[1];
-    const validFileByteSize = "30000";
-    const fileSize = file.size;
-
-    return {
-      fileExtension: validExtensions.includes(fileExtensions),
-      fileSizeValid: fileSize <= validFileByteSize,
-    };
-  };
-
   const handleSubmit = async values => {
     const formData = new FormData();
 
@@ -166,68 +105,6 @@ const ClientForm = ({
     }
   };
 
-  const LogoComponent = () => (
-    <div className="my-4 flex flex-row">
-      <div className="mt-2 h-30 w-30 border border-dashed border-miru-dark-purple-400">
-        <div className="profile-img relative m-auto h-30 w-30 cursor-pointer text-center text-xs font-semibold">
-          <Avatar
-            classNameImg="h-full w-full md:h-full md:w-full"
-            url={clientLogoUrl}
-          />
-          <div className="hover-edit absolute top-0 left-0 h-full w-full bg-miru-white-1000 p-4 opacity-80">
-            <button
-              className="flex flex-row text-miru-han-purple-1000"
-              type="button"
-            >
-              <label className="flex cursor-pointer" htmlFor="file_input">
-                <img
-                  alt="edit"
-                  className="cursor-pointer rounded-full"
-                  src={EditImageButtonSVG}
-                  style={{ minWidth: "40px" }}
-                />
-                <p className="my-auto">Edit</p>
-              </label>
-              <input
-                className="hidden"
-                id="file_input"
-                name="logo"
-                type="file"
-                onChange={onLogoChange}
-              />
-            </button>
-            {clientLogoUrl && (
-              <button
-                className="flex flex-row pl-2 text-miru-red-400"
-                type="button"
-                onClick={handleDeleteLogo}
-              >
-                <img
-                  alt="delete"
-                  src={deleteImageIcon}
-                  style={{ minWidth: "20px" }}
-                />
-                <p className="pl-3">Delete</p>
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-      <div className="my-auto ml-6 text-xs font-normal text-miru-dark-purple-400">
-        <p>Accepted file formats: PNG and JPG.</p>
-        <p>File size should be &#8826; 30 KB.</p>
-        <p>Image resolution should be 1:1.</p>
-      </div>
-      <input
-        className="hidden"
-        id="file_input"
-        name="logo"
-        type="file"
-        onClick={onLogoChange}
-      />
-    </div>
-  );
-
   return (
     <Formik
       initialValues={getInitialvalues(clientData)}
@@ -243,38 +120,14 @@ const ClientForm = ({
               <div className="my-4">
                 <div className="field">
                   <div className="mt-1">
-                    {formType == "edit" ? (
-                      <LogoComponent />
-                    ) : clientLogoUrl ? (
-                      <LogoComponent />
-                    ) : (
-                      <div className="mt-2 flex flex-row">
-                        <div className="mt-2 h-30 w-30 border border-dashed border-miru-dark-purple-400 ">
-                          <label
-                            className="flex h-full w-full cursor-pointer justify-center"
-                            htmlFor="file-input"
-                          >
-                            <div className="m-auto cursor-pointer text-center text-xs font-semibold">
-                              <p className="text-miru-han-purple-1000">
-                                Select File
-                              </p>
-                            </div>
-                          </label>
-                          <input
-                            className="hidden"
-                            id="file-input"
-                            name="logo"
-                            type="file"
-                            onChange={onLogoChange}
-                          />
-                        </div>
-                        <div className="my-auto ml-6 text-xs font-normal text-miru-dark-purple-400">
-                          <p>Accepted file formats: PNG and JPG.</p>
-                          <p>File size should be &#8826; 30 KB.</p>
-                          <p>Image resolution should be 1:1.</p>
-                        </div>
-                      </div>
-                    )}
+                    <UploadLogo
+                      clientLogoUrl={clientLogoUrl}
+                      formType={formType}
+                      handleDeleteLogo={handleDeleteLogo}
+                      setClientLogo={setClientLogo}
+                      setClientLogoUrl={setClientLogoUrl}
+                      setFileUploadError={setFileUploadError}
+                    />
                     <p className="mt-3 block max-w-xs text-center text-xs tracking-wider text-red-600">
                       {fileUploadError}
                     </p>
@@ -450,5 +303,32 @@ const ClientForm = ({
     </Formik>
   );
 };
+
+interface IClientForm {
+  clientLogoUrl: string;
+  handleDeleteLogo: any;
+  setClientLogoUrl: any;
+  setClientLogo: any;
+  formType?: string;
+  clientData?: any;
+  setClientData?: any;
+  setnewClient?: any;
+  clientLogo?: any;
+  setApiError?: any;
+  setShowEditDialog?: any;
+}
+
+interface FormValues {
+  name: string;
+  email: string;
+  phone: string;
+  address1: string;
+  address2: string;
+  country: any;
+  state: any;
+  city: any;
+  zipcode: string;
+  logo: any;
+}
 
 export default ClientForm;
